requirements.txt
aiogram==3.1.1          # Telegram Bot API
openai==1.3.5           # OpenRouter client  
python-dotenv==1.0.0    # Environment variables
